Refactors directory structure.
