+++ Backend/app/file_paths.py
@@ -1,0 +1,7 @@
+from pathlib import Path
+
+DATA_DIR = Path(".") / "app" / "ai" / "data"
+TOPICS_FILE = DATA_DIR / "topics.json"
+MODEL_PATH = DATA_DIR / "skynet.h5"
+WORDS_FILE = DATA_DIR / "words.pkl"
+TOPIC_TYPES_FILE = DATA_DIR / "topic_types.pkl"
